[front] refactor: simplify Poke trigger status route
